+++ groups/item_sites_item_term_store_groups_item_sets_item_relations_request_builder.go
@@ -29,9 +29,9 @@ type ItemSitesItemTermStoreGroupsItemSetsItemRelationsRequestBuilderGetQueryPara
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int64 `uriparametername:"%24skip"`
+    Skip *int32 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int64 `uriparametername:"%24top"`
+    Top *int32 `uriparametername:"%24top"`
 }
 // ItemSitesItemTermStoreGroupsItemSetsItemRelationsRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type ItemSitesItemTermStoreGroupsItemSetsItemRelationsRequestBuilderGetRequestConfiguration struct {
